documentation-requirements.txt
Sphinx==1.8.5
sphinx-autobuild==0.7.1
sphinx-rtd-theme==0.4.3
sphinxcontrib-websupport==1.1.2
numpydoc==0.9.1
